Release ATK 2.29.94 (development)
This is a special development release to include all the changes so
far on master. But ATK 2.30 will be based on 2.29.92 plus the revert
of commit "atkimplementor: use the G_DEFINE_INTERFACE macro to declare
it as interface in the introspection", as the other two commits were
not tested yet.