meson: Fix usage of glib as a subproject dependency
We include glib headers, so we need to ensure that we also pull in the
glib dependency correctly, to inherit the right include flags. Also
doesn't lead to overlinking because meson uses --as-needed by default.

This wasn't noticed earlier because when building against glib inside
a prefix, glib and gobject headers are in the same location.